Create zotero-desktop
I added the Zotero Reference Manager to winapps.
Zotero is a free software that manages scientific references in word files. 
It makes citing scientific papers much easier.

The icon I used isn't pretty, but it gets the job done. Also, the icon is free to use (as Zotero states in their own website). 

For more references,
https://www.zotero.org/